+++ plugins/cong_avoidance/dtcp-ps-cas.c
@@ -28,7 +28,6 @@
 #include "rds/rmem.h"
 #include "dtcp-ps.h"
 #include "dtcp-conf-utils.h"
-#include "dtcp-ps-common.h"
 #include "policies.h"
 #include "logs.h"
 
@@ -61,25 +60,6 @@ struct cas_dtcp_ps_data {
 	int         ws_index;
 #endif
 };
-
-static int
-cas_lost_control_pdu(struct dtcp_ps * ps)
-{ return common_lost_control_pdu(ps); }
-
-static int cas_rcvr_ack(struct dtcp_ps * ps, const struct pci * pci)
-{ return common_rcvr_ack(ps, pci); }
-
-static int
-cas_sender_ack(struct dtcp_ps * ps, seq_num_t seq_num)
-{ return common_sender_ack(ps, seq_num); }
-
-static int
-cas_sending_ack(struct dtcp_ps * ps, seq_num_t seq)
-{ return common_sending_ack(ps, seq); }
-
-static int
-cas_receiving_flow_control(struct dtcp_ps * ps, const struct pci * pci)
-{ return common_receiving_flow_control(ps, pci); }
 
 static unsigned int round_half_to_even(unsigned int real_window)
 {
@@ -171,10 +151,6 @@ cas_rcvr_flow_control(struct dtcp_ps * ps, const struct pci * pci)
         return 0;
 }
 
-static int
-cas_rate_reduction(struct dtcp_ps * ps)
-{ return common_rate_reduction(ps); }
-
 static int dtcp_ps_cas_set_policy_set_param(struct ps_base * bps,
                                             const char    * name,
                                             const char    * value)
@@ -206,9 +182,6 @@ static int dtcp_ps_cas_set_policy_set_param(struct ps_base * bps,
         }
         return 0;
 }
-
-static int cas_rtt_estimator(struct dtcp_ps * ps, seq_num_t sn)
-{ return common_rtt_estimator(ps, sn); }
 
 static struct ps_base *
 dtcp_ps_cas_create(struct rina_component * component)
@@ -262,20 +235,20 @@ dtcp_ps_cas_create(struct rina_component * component)
                                          policy_param_value(ps_param));
 
         ps->flow_init                   = NULL;
-        ps->lost_control_pdu            = cas_lost_control_pdu;
-        ps->rtt_estimator               = cas_rtt_estimator;
+        ps->lost_control_pdu            = NULL; /* default */
+        ps->rtt_estimator               = NULL; /* default */
         ps->retransmission_timer_expiry = NULL;
         ps->received_retransmission     = NULL;
-        ps->sender_ack                  = cas_sender_ack;
-        ps->sending_ack                 = cas_sending_ack;
+        ps->sender_ack                  = NULL; /* default */
+        ps->sending_ack                 = NULL; /* default */
         ps->receiving_ack_list          = NULL;
         ps->initial_rate                = NULL;
-        ps->receiving_flow_control      = cas_receiving_flow_control;
+        ps->receiving_flow_control      = NULL; /* default */
         ps->update_credit               = NULL;
         ps->reconcile_flow_conflict     = NULL;
-        ps->rcvr_ack                    = cas_rcvr_ack,
+        ps->rcvr_ack                    = NULL; /* default */
         ps->rcvr_flow_control           = cas_rcvr_flow_control;
-        ps->rate_reduction              = cas_rate_reduction;
+        ps->rate_reduction              = NULL; /* default */
         ps->rcvr_control_ack            = NULL;
         ps->no_rate_slow_down           = NULL;
         ps->no_override_default_peak    = NULL;
